Refactored And Fixed Results Edit/Display
