+++ client/src/constants.ts
@@ -7,7 +7,7 @@ export const token = localStorage.getItem('accessToken')
 export const maximumEntries = 10
 export const maximumEntriesEE = 5
 export const storedUserID = localStorage.getItem('userID')
-export const header = ['計画	']
+export const header = ['計画']
 export const smallDate = ['2022/24月', '2022/25月', '2022/26月']
 export const dates = ['04月', '05月', '06月', '07月', '08月', '09月', '10月', '11月', '12月', '1月', '2月', '3月']
 export const months = [4, 5, 6, 7, 8, 9, 10, 11, 12, 1, 2, 3]
