+++ client/src/pages/Results/resultsDownloadXLS.tsx
@@ -1,0 +1,650 @@
+import { getResultsA } from "../../api/ResultsEndpoint/GetResultsA"
+import { ACCESS_TOKEN, monthNames, months } from "../../constants"
+import { translate } from "../../utils/translationUtil"
+import * as XLSX from 'xlsx'
+
+
+export const downloadXLS = (language) => {
+
+  const additionalHeaders = {
+    1: { en: 'H1', jp: '上期計	' },
+    2: { en: 'H2', jp: '下期計	' },
+    3: { en: 'Year Total', jp: '合計' },
+    4: { en: 'Sales%', jp: '売上比' },
+  }
+  
+  getResultsA(localStorage.getItem(ACCESS_TOKEN))
+    .then((response) => {
+      const aggregatedData = response.cost_of_sales_results.reduce((acc, item) => {
+        const { month, ...values } = item
+        if (!acc[month]) {
+          acc[month] = { ...values }
+        } else {
+          Object.keys(values).forEach((key) => {
+            acc[month][key] += values[key]
+          })
+        }
+        return acc
+      }, {})
+      const aggregatedExpensesData = response.expenses_results.reduce((acc, item) => {
+        const { month, ...values } = item
+        if (!acc[month]) {
+          acc[month] = { month, ...values } // Include month in the object
+        } else {
+          Object.keys(values).forEach((key) => {
+            acc[month][key] += values[key]
+          })
+        }
+        return acc
+      }, {})
+
+      const aggregateEmployeeResultsData = (employeesResults) => {
+        const aggregatedData = {}
+        let totalAnnualExecutive = 0
+        let totalAnnualSalary = 0
+        let totalBonusAndFuelAllowance = 0
+        let totalWelfareExpense = 0
+        let totalStatutoryWelfareExpense = 0
+        let totalInsurancePremium = 0
+
+        employeesResults.forEach((employee) => {
+          totalAnnualExecutive += employee.executive_remuneration
+          totalAnnualSalary += employee.salary
+          totalBonusAndFuelAllowance += employee.bonus_and_fuel_allowance
+          totalWelfareExpense += parseFloat(employee.welfare_expense)
+          totalStatutoryWelfareExpense += parseFloat(employee.statutory_welfare_expense)
+          totalInsurancePremium += parseFloat(employee.insurance_premium)
+        })
+
+        const monthlyExecutive = totalAnnualExecutive / 12
+        const monthlySalary = totalAnnualSalary / 12
+        const yearlyBonusAndFuelAllowance = totalBonusAndFuelAllowance
+        const monthlyWelfareExpense = (totalAnnualSalary * 0.0048) / 12
+        const monthStatutoryWelfareExpense = (totalAnnualSalary * 0.0048) / 12
+        const monthlyInsurancePremium = totalInsurancePremium / 12
+
+        months.forEach((month) => {
+          aggregatedData[month] = {
+            executive_remuneration: monthlyExecutive,
+            salary: monthlySalary,
+            bonus_and_fuel_allowance: yearlyBonusAndFuelAllowance,
+            welfare_expense: monthlyWelfareExpense,
+            statutory_welfare_expense: monthStatutoryWelfareExpense,
+            insurance_premium: monthlyInsurancePremium,
+          }
+        })
+
+        return aggregatedData
+      }
+
+      const aggregatedEmployeeExpensesResults = response.employee_expenses_results.reduce((acc, item) => {
+        const { month, employee, project, ...values } = item
+
+        if (!acc[month]) {
+          acc[month] = {
+            month,
+            employees: [employee],
+            projects: [project],
+            totalSalary: Number(employee.salary) || 0,
+            ...values,
+          }
+        } else {
+          acc[month].employees.push(employee)
+          acc[month].projects.push(project)
+
+          acc[month].totalSalary += Number(employee.salary) || 0
+
+          Object.keys(values).forEach((key) => {
+            if (typeof values[key] === 'number') {
+              acc[month][key] += values[key]
+            } else if (typeof values[key] === 'string') {
+              acc[month][key] = values[key]
+            }
+          })
+        }
+        return acc
+      }, {})
+
+      const aggregatedProjectSalesResultsData = response.project_sales_results.reduce((acc, item) => {
+        const { project, ...values } = item
+        const month = project?.month
+        if (!month) {
+          return acc
+        }
+        if (!acc[month]) {
+          acc[month] = { month }
+        }
+        Object.keys(values).forEach((key) => {
+          const value = parseFloat(values[key])
+          if (!isNaN(value)) {
+            acc[month][key] = (acc[month][key] || 0) + value
+          }
+        })
+
+        return acc
+      }, {})
+
+      // SALES REVENUE
+      const salesValues = months.map((month) => aggregatedProjectSalesResultsData[month]?.sales_revenue || 0)
+
+      //COST OF SALES
+      const costOfSalesValues = months.map((month) => {
+        const purchases = aggregatedData[month]?.purchase || 0
+        const outsourcing = aggregatedData[month]?.outsourcing_expense || 0
+        const productPurchase = aggregatedData[month]?.product_purchase || 0
+        const dispatchLabor = aggregatedData[month]?.dispatch_labor_expense || 0
+        const communicationCost = aggregatedData[month]?.communication_expense || 0
+        const workInProgress = aggregatedData[month]?.work_in_progress_expense || 0
+        const amortization = aggregatedData[month]?.amortization_expense || 0
+        return (
+          purchases + outsourcing + productPurchase + dispatchLabor + communicationCost + workInProgress + amortization
+        )
+      })
+      const purchasesValues = months.map((month) => aggregatedData[month]?.purchase || 0)
+      const outsourcingExpenseValues = months.map((month) => aggregatedData[month]?.outsourcing_expense || 0)
+      const productPurchaseValues = months.map((month) => aggregatedData[month]?.product_purchase || 0)
+      const dispatchLaborExpenseValues = months.map((month) => aggregatedData[month]?.dispatch_labor_expense || 0)
+      const communicationCostValues = months.map((month) => aggregatedData[month]?.communication_expense || 0)
+      const workInProgressValues = months.map((month) => aggregatedData[month]?.work_in_progress_expense || 0)
+      const amortizationValues = months.map((month) => aggregatedData[month]?.amortization_expense || 0)
+
+      // GROSS PROFIT
+      const grossProfitValues = months.map((month, index) => {
+        const totalSales = salesValues[index]
+        const totalCostOfSales = costOfSalesValues[index]
+        const grossProfit = totalSales - totalCostOfSales
+        return grossProfit
+      })
+
+      // EMPLOYEE EXPENSE
+      const employeeExpensesValues = months.map((month) => {
+        const executiveRemuneration = aggregatedExpensesData[month]?.executive_remuneration || 0
+        const salary = aggregatedEmployeeExpensesResults[month]?.totalSalary || 0
+        const fuel_allowance = aggregatedExpensesData[month]?.fuel_allowance || 0
+        const statutory_welfare_expense = aggregatedExpensesData[month]?.statutory_welfare_expense || 0
+        const welfare_expense = aggregatedExpensesData[month]?.welfare_expense || 0
+        const insurance_premiums = aggregatedExpensesData[month]?.insurance_premiums || 0
+
+        return (
+          executiveRemuneration +
+          salary +
+          fuel_allowance +
+          statutory_welfare_expense +
+          welfare_expense +
+          insurance_premiums
+        )
+      })
+      // EMPLOYEES
+      const result = aggregateEmployeeResultsData(response.employees_results)
+      const executiveRemunerationValues = months.map((month) => result[month]?.executive_remuneration || 0)
+      const salaryValues = months.map((month) => result[month]?.salary || 0)
+      const totalBonusAndFuelAllowance = result[12]?.bonus_and_fuel_allowance || 0
+      const bonusAndFuelAllowanceValues = months.map((month) => {
+        return month === 12 ? totalBonusAndFuelAllowance : 0
+      })
+      const statutoryWelfareExpenseValues = months.map((month) => result[month]?.statutory_welfare_expense || 0)
+      const welfareExpenseValues = months.map((month) => result[month]?.welfare_expense || 0)
+      const insurancePremiumsValues = months.map((month) => result[month]?.insurance_premium || 0)
+
+      // EXPENSES
+      const expenseValues = months.map((month) => {
+        const consumables = aggregatedExpensesData[month]?.consumable_expense || 0
+        const rent = aggregatedExpensesData[month]?.rent_expense || 0
+        const taxAndPublicCharge = aggregatedExpensesData[month]?.tax_and_public_charge || 0
+        const depreciation = aggregatedExpensesData[month]?.depreciation_expense || 0
+        const travel_expense = aggregatedExpensesData[month]?.travel_expense || 0
+        const communication_expense = aggregatedExpensesData[month]?.communication_expense || 0
+        const utilities_expense = aggregatedExpensesData[month]?.utilities_expense || 0
+        const transaction_fee = aggregatedExpensesData[month]?.transaction_fee || 0
+        const advertising_expense = aggregatedExpensesData[month]?.advertising_expense || 0
+        const entertainment_expense = aggregatedExpensesData[month]?.entertainment_expense || 0
+        const professional_service_fee = aggregatedExpensesData[month]?.professional_service_fee || 0
+        return (
+          consumables +
+          rent +
+          taxAndPublicCharge +
+          depreciation +
+          travel_expense +
+          communication_expense +
+          utilities_expense +
+          transaction_fee +
+          advertising_expense +
+          entertainment_expense +
+          professional_service_fee
+        )
+      })
+      const consumableValues = months.map((month) => aggregatedExpensesData[month]?.consumable_expense || 0)
+      const rentValues = months.map((month) => aggregatedExpensesData[month]?.rent_expense || 0)
+      const taxesPublicChargesValues = months.map((month) => aggregatedExpensesData[month]?.tax_and_public_charge || 0)
+      const depreciationExpensesValues = months.map((month) => aggregatedExpensesData[month]?.depreciation_expense || 0)
+      const travelExpenseValues = months.map((month) => aggregatedExpensesData[month]?.travel_expense || 0)
+      const communicationExpenseValues = months.map(
+        (month) => aggregatedExpensesData[month]?.communication_expense || 0,
+      )
+      const utilitiesValues = months.map((month) => aggregatedExpensesData[month]?.utilities_expense || 0)
+      const transactionFeeValues = months.map((month) => aggregatedExpensesData[month]?.transaction_fee || 0)
+      const advertisingExpenseValues = months.map((month) => aggregatedExpensesData[month]?.advertising_expense || 0)
+      const entertainmentExpenseValues = months.map(
+        (month) => aggregatedExpensesData[month]?.entertainment_expense || 0,
+      )
+      const professionalServiceFeeValues = months.map(
+        (month) => aggregatedExpensesData[month]?.professional_service_fee || 0,
+      )
+
+      // SELLING AND GENERAL ADMIN EXPENSES
+      const sellingAndGeneralAdminExpenseValues = months.map((month, index) => {
+        const total_employee_expense = employeeExpensesValues[index]
+        const total_expense = expenseValues[index]
+        const sellingAndGeneralAdminExpense = total_employee_expense + total_expense
+        return sellingAndGeneralAdminExpense
+      })
+
+      // OPERATING INCOME
+      const operatingIncomeValues = months.map((month, index) => {
+        const gross_profit = grossProfitValues[index]
+        const selling_and_general_admin = sellingAndGeneralAdminExpenseValues[index]
+        const operating_income_value = gross_profit - selling_and_general_admin
+        return operating_income_value
+      })
+      //NoN Operating Income & Expense
+      const nonOperatingIncomeValues = months.map(
+        (month) => aggregatedProjectSalesResultsData[month]?.non_operating_income || 0,
+      )
+      const nonOperatingExpensesValues = months.map(
+        (month) => aggregatedProjectSalesResultsData[month]?.non_operating_expense || 0,
+      )
+
+      const ordinaryProfitValues = months.map((month, index) => {
+        const operating_income = operatingIncomeValues[index]
+        const non_operating_income = nonOperatingIncomeValues[index]
+        const totalOperating = operating_income + non_operating_income
+        const totalOrdinaryIncome = totalOperating - nonOperatingExpensesValues[index]
+
+        return totalOrdinaryIncome
+      })
+
+      const cumulativeSum = (arr) => {
+        let sum = 0
+        return arr.map((value) => (sum += value))
+      }
+      const cumulativeOrdinaryProfitValues = cumulativeSum(ordinaryProfitValues)
+
+      const firstHalfTotal = (arr) => arr.slice(0, 6).reduce((acc, value) => acc + parseFloat(value), 0)
+      const secondHalfTotal = (arr) => arr.slice(6).reduce((acc, value) => acc + parseFloat(value), 0)
+      const total = (arr) => arr.reduce((acc, value) => acc + parseFloat(value), 0)
+      const header = translate('results', language)
+      const data = [
+        {
+          label: translate('salesRevenue', language),
+          values: [...salesValues, firstHalfTotal(salesValues), secondHalfTotal(salesValues), total(salesValues), '0'],
+        },
+        {
+          label: translate('sales', language),
+          values: [...salesValues, firstHalfTotal(salesValues), secondHalfTotal(salesValues), total(salesValues), '0'],
+        },
+        //start of cost of sales portion
+        {
+          label: translate('costOfSales', language),
+          values: [
+            ...costOfSalesValues,
+            firstHalfTotal(costOfSalesValues),
+            secondHalfTotal(costOfSalesValues),
+            total(costOfSalesValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('purchases', language),
+          values: [
+            ...purchasesValues,
+            firstHalfTotal(purchasesValues),
+            secondHalfTotal(purchasesValues),
+            total(purchasesValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('outsourcingExpenses', language),
+          values: [
+            ...outsourcingExpenseValues,
+            firstHalfTotal(outsourcingExpenseValues),
+            secondHalfTotal(outsourcingExpenseValues),
+            total(outsourcingExpenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('productPurchases', language),
+          values: [
+            ...productPurchaseValues,
+            firstHalfTotal(productPurchaseValues),
+            secondHalfTotal(productPurchaseValues),
+            total(productPurchaseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('dispatchLaborExpenses', language),
+          values: [
+            ...dispatchLaborExpenseValues,
+            firstHalfTotal(dispatchLaborExpenseValues),
+            secondHalfTotal(dispatchLaborExpenseValues),
+            total(dispatchLaborExpenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('communicationExpenses', language),
+          values: [
+            ...communicationCostValues,
+            firstHalfTotal(communicationCostValues),
+            secondHalfTotal(communicationCostValues),
+            total(communicationCostValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('workInProgressExpenses', language),
+          values: [
+            ...workInProgressValues,
+            firstHalfTotal(workInProgressValues),
+            secondHalfTotal(workInProgressValues),
+            total(workInProgressValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('amortizationExpenses', language),
+          values: [
+            ...amortizationValues,
+            firstHalfTotal(amortizationValues),
+            secondHalfTotal(amortizationValues),
+            total(amortizationValues),
+            '0',
+          ],
+        },
+        // end for cost of sales section
+        {
+          label: translate('grossProfit', language),
+          values: [
+            ...grossProfitValues,
+            firstHalfTotal(grossProfitValues),
+            secondHalfTotal(grossProfitValues),
+            total(grossProfitValues),
+            '',
+          ],
+        },
+        // start for employee expense section
+        {
+          label: translate('employeeExpenses', language),
+          values: [
+            ...employeeExpensesValues,
+            firstHalfTotal(employeeExpensesValues),
+            secondHalfTotal(employeeExpensesValues),
+            total(employeeExpensesValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('executiveRemunerations', language),
+          values: [
+            ...executiveRemunerationValues,
+            firstHalfTotal(executiveRemunerationValues),
+            secondHalfTotal(executiveRemunerationValues),
+            total(executiveRemunerationValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('salary', language),
+          values: [
+            ...salaryValues,
+            firstHalfTotal(salaryValues),
+            secondHalfTotal(salaryValues),
+            total(salaryValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('bonusAndFuelAllowances', language),
+          values: [
+            ...bonusAndFuelAllowanceValues,
+            firstHalfTotal(bonusAndFuelAllowanceValues),
+            secondHalfTotal(bonusAndFuelAllowanceValues),
+            total(bonusAndFuelAllowanceValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('statutoryWelfareExpenses', language),
+          values: [
+            ...statutoryWelfareExpenseValues,
+            firstHalfTotal(statutoryWelfareExpenseValues),
+            secondHalfTotal(statutoryWelfareExpenseValues),
+            total(statutoryWelfareExpenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('welfareExpenses', language),
+          values: [
+            ...welfareExpenseValues,
+            firstHalfTotal(welfareExpenseValues),
+            secondHalfTotal(welfareExpenseValues),
+            total(welfareExpenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('insurancePremiums', language),
+          values: [
+            ...insurancePremiumsValues,
+            firstHalfTotal(insurancePremiumsValues),
+            secondHalfTotal(insurancePremiumsValues),
+            total(insurancePremiumsValues),
+            '0',
+          ],
+        },
+        //end for employee expense section
+        //start for expenses section
+        {
+          label: translate('expenses', language),
+          values: [
+            ...expenseValues,
+            firstHalfTotal(expenseValues),
+            secondHalfTotal(expenseValues),
+            total(expenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('consumableExpenses', language),
+          values: [
+            ...consumableValues,
+            firstHalfTotal(consumableValues),
+            secondHalfTotal(consumableValues),
+            total(consumableValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('rentExpenses', language),
+          values: [...rentValues, firstHalfTotal(rentValues), secondHalfTotal(rentValues), total(rentValues), '0'],
+        },
+        {
+          label: translate('taxesAndPublicCharges', language),
+          values: [
+            ...taxesPublicChargesValues,
+            firstHalfTotal(taxesPublicChargesValues),
+            secondHalfTotal(taxesPublicChargesValues),
+            total(taxesPublicChargesValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('depreciationExpenses', language),
+          values: [
+            ...depreciationExpensesValues,
+            firstHalfTotal(depreciationExpensesValues),
+            secondHalfTotal(depreciationExpensesValues),
+            total(depreciationExpensesValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('travelExpenses', language),
+          values: [
+            ...travelExpenseValues,
+            firstHalfTotal(travelExpenseValues),
+            secondHalfTotal(travelExpenseValues),
+            total(travelExpenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('communicationExpenses', language),
+          values: [
+            ...communicationExpenseValues,
+            firstHalfTotal(communicationExpenseValues),
+            secondHalfTotal(communicationExpenseValues),
+            total(communicationExpenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('utilitiesExpenses', language),
+          values: [
+            ...utilitiesValues,
+            firstHalfTotal(utilitiesValues),
+            secondHalfTotal(utilitiesValues),
+            total(utilitiesValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('transactionFees', language),
+          values: [
+            ...transactionFeeValues,
+            firstHalfTotal(transactionFeeValues),
+            secondHalfTotal(transactionFeeValues),
+            total(transactionFeeValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('advertisingExpenses', language),
+          values: [
+            ...advertisingExpenseValues,
+            firstHalfTotal(advertisingExpenseValues),
+            secondHalfTotal(advertisingExpenseValues),
+            total(advertisingExpenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('entertainmentExpenses', language),
+          values: [
+            ...entertainmentExpenseValues,
+            firstHalfTotal(entertainmentExpenseValues),
+            secondHalfTotal(entertainmentExpenseValues),
+            total(entertainmentExpenseValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('professionalServicesFees', language),
+          values: [
+            ...professionalServiceFeeValues,
+            firstHalfTotal(professionalServiceFeeValues),
+            secondHalfTotal(professionalServiceFeeValues),
+            total(professionalServiceFeeValues),
+            '0',
+          ],
+        },
+        // end for expense section
+        {
+          //add 人件費 + 経費 field
+          label: translate('sellingAndGeneralAdminExpenses', language), // shortened version as it is too long in English Mode translate('sellingAndGeneralAdminExpenses', language)
+          values: [
+            ...sellingAndGeneralAdminExpenseValues,
+            firstHalfTotal(sellingAndGeneralAdminExpenseValues),
+            secondHalfTotal(sellingAndGeneralAdminExpenseValues),
+            total(sellingAndGeneralAdminExpenseValues),
+            '0',
+          ],
+        },
+        //Operating income 営業利益 ①
+        {
+          label: translate('operatingIncome', language),
+          values: [
+            ...operatingIncomeValues,
+            firstHalfTotal(operatingIncomeValues),
+            secondHalfTotal(operatingIncomeValues),
+            total(operatingIncomeValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('nonOperatingIncome', language),
+          values: [
+            ...nonOperatingIncomeValues,
+            firstHalfTotal(nonOperatingIncomeValues),
+            secondHalfTotal(nonOperatingIncomeValues),
+            total(nonOperatingIncomeValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('nonOperatingExpenses', language),
+          values: [
+            ...nonOperatingExpensesValues,
+            firstHalfTotal(nonOperatingExpensesValues),
+            secondHalfTotal(nonOperatingExpensesValues),
+            total(nonOperatingExpensesValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('ordinaryIncome', language),
+          values: [
+            ...ordinaryProfitValues,
+            firstHalfTotal(ordinaryProfitValues),
+            secondHalfTotal(ordinaryProfitValues),
+            total(ordinaryProfitValues),
+            '0',
+          ],
+        },
+        {
+          label: translate('cumulativeOrdinaryIncome', language),
+          values: [
+            ...cumulativeOrdinaryProfitValues,
+            firstHalfTotal(cumulativeOrdinaryProfitValues),
+            secondHalfTotal(cumulativeOrdinaryProfitValues),
+            total(cumulativeOrdinaryProfitValues),
+            '0',
+          ],
+        },
+      ]
+      // Updated translation base on what language user has selected.
+      const monthNamesTranslation = Object.values(monthNames).map((month) => month[language])
+      const reorderedMonthNames = [...monthNamesTranslation.slice(3), ...monthNamesTranslation.slice(0, 3)] //Changed the order of months [Apr - Mar]
+      const additionalHeadersTranslation = Object.values(additionalHeaders).map((month) => month[language])
+      const excelRows = [
+        ['', ...reorderedMonthNames, ...additionalHeadersTranslation],
+        ['', header, header, header, header, ...Array(11).fill(header)],
+      ]
+
+      data.forEach((item) => {
+        excelRows.push([item.label, ...item.values])
+      })
+      // This part of the code will write the excel
+      const worksheet = XLSX.utils.aoa_to_sheet(excelRows)
+      const workbook = XLSX.utils.book_new()
+      const resultsSummary = translate('resultsSummary', language)
+      XLSX.utils.book_append_sheet(workbook, worksheet, translate('resultsSummary', language))
+      XLSX.writeFile(workbook, `${resultsSummary}.xlsx`)
+    })
+    .catch((error) => {
+      console.log(error)
+    })
+}
