+++ client/src/reducers/costOfSale/costOfSaleResultSlice.ts
@@ -1,6 +1,7 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
 import { CostOfSaleResultEntity } from '../../entity/costOfSaleResultEntity'
 import { fetchWithPolling } from '../../utils/helperFunctionsUtil'
+import { mappingCostOfSalesLabels } from '../../utils/labelConverter'
 
 const initialState = {
   isLoading: false,
@@ -8,6 +9,7 @@ const initialState = {
   monthlyTotals: [],
   yearlyTotal: 0,
   monthlyTotalsByDate: [],
+  editableData: [],
 }
 
 export const fetchCostOfSaleResult = createAsyncThunk('cost-of-sale-result/fetch', async () => {
@@ -18,6 +20,23 @@ const costOfSaleResult = createSlice({
   name: 'costOfSaleResult',
   initialState,
   reducers: {
+    updateCostOfSalesResultsScreen: (state, action) => {
+      state.list = state.list.map((item) => {
+        const recordToBeChanged = action.payload.modifiedFields[item.cost_of_sale_result_id]
+
+        if (recordToBeChanged) {
+          const updatedItem = { ...item }
+
+          recordToBeChanged.forEach((record) => {
+            const recordLabel = record.label
+            const convertedLabel = mappingCostOfSalesLabels(recordLabel)
+            updatedItem[convertedLabel] = record.value
+          })
+          return updatedItem
+        }
+        return item // Return item as is if no match
+      })
+    },
   },
   extraReducers(builder) {
     builder
@@ -33,5 +52,7 @@ const costOfSaleResult = createSlice({
       })
   },
 })
+
+export const { updateCostOfSalesResultsScreen } = costOfSaleResult.actions
 
 export default costOfSaleResult.reducer
