+++ client/src/reducers/expenses/expensesResultsSlice.ts
@@ -1,6 +1,7 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
 import { fetchWithPolling } from '../../utils/helperFunctionsUtil'
 import { ExpenseResultEntity } from '../../entity/expenseResultEntity'
+import { mappingExpensesLabels } from '../../utils/labelConverter'
 
 const initialState = {
   isLoading: false,
@@ -18,6 +19,23 @@ const expenseResults = createSlice({
   name: 'expenseResults',
   initialState,
   reducers: {
+    updateExpensesResultsScreen: (state, action) => {
+      state.list = state.list.map((item) => {
+        const recordToBeChanged = action.payload.modifiedFields[item.expense_result_id]
+
+        if (recordToBeChanged) {
+          const updatedItem = { ...item }
+
+          recordToBeChanged.forEach((record) => {
+            const recordLabel = record.label
+            const convertedLabel = mappingExpensesLabels(recordLabel)
+            updatedItem[convertedLabel] = record.value
+          })
+          return updatedItem
+        }
+        return item // Return item as is if no match
+      })
+    },
   },
   extraReducers(builder) {
     builder
@@ -33,5 +51,7 @@ const expenseResults = createSlice({
       })
   },
 })
+
+export const { updateExpensesResultsScreen } = expenseResults.actions
 
 export default expenseResults.reducer
