+++ client/src/reducers/expenses/expensesSlice.ts
@@ -21,7 +21,6 @@ const expense = createSlice({
   initialState,
   reducers: {
     updateExpensesPlanningScreen: (state, action) => {
-      console.log("inside expense",action.payload)
       state.list = state.list.map((item) => {
         const recordToBeChanged = action.payload.modifiedFields[item.expense_id]
